Chief-Decifer-recovered/requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-draggable==0.1.2
plotly==5.22.0
pandas==2.2.2
python-dotenv==1.0.1
anthropic>=0.25.0
yfinance>=0.2.40
lxml>=5.0.0
# Sentiment scoring pipeline (FinBERT local scorer)
# Pinned: torch 2.6+ requires Python 3.10+; use 2.2.x on Python 3.9
# transformers 4.51+ blocks torch.load on torch<2.6; pin to 4.48.x
# numpy 2.x is incompatible with torch 2.2.x binaries; pin to <2.0
transformers==4.48.3
torch==2.2.2
numpy<2.0
safetensors
